refactor: create new component for mode switcher
