+++ packages/apps/intervention/src/components/organisms/index.ts
@@ -20,6 +20,7 @@ export {default as ClientProspectSearchBar} from './ClientProspectSearchBar/Clie
 export {default as ContractSearchBar} from './ContractSearchBar/ContractSearchBar';
 export {default as DropdownCards} from './DropdownCards/DropdownCards';
 export {default as EquipmentFamilySearchBar} from './EquipmentFamilySearchBar/EquipmentFamilySearchBar';
+export {default as EquipmentModeSwitch} from './EquipmentModeSwitch/EquipmentModeSwitch';
 export {default as InterventionNoteTypePicker} from './InterventionNoteTypePicker/InterventionNoteTypePicker';
 export {default as PlaceEquipmentSearchBar} from './PlaceEquipmentSearchBar/PlaceEquipmentSearchBar';
 export {default as TypePicker} from './TypePicker/TypePicker';
